[Move] CometReflectionTypeLoadException to exceptions folder
